requirements.txt
pyTelegramBotAPI==4.14.1
deep-translator==1.11.4
gtts==2.5.1
python-dotenv
